Add sweep1.c PPE latency sweeper
This program is identical in functionality to sweep.c, but highlights an
interesting set of behaviors out of BOOM. The hot loop's disassembly is
shown below.

```
   10714:       |  |  |  /----------> c0002673                  rdcycle a2
   10718:       |  |  |  |  /-------- c085                      beqz    s1,10738 <main+0x98>
   1071a:       |  |  |  |  |         4781                      li      a5,0
   1071c:       |  |  |  |  |  /----- a031                      j       10728 <main+0x88>
   1071e:       |  |  |  |  |  |  /-> 1af7f6d3                  fdiv.d  fa3,fa5,fa5
   10722:       |  |  |  |  |  |  |   2785                      addiw   a5,a5,1
   10724:       |  |  |  |  +--|--|-- 00f48a63                  beq     s1,a5,10738 <main+0x98>
   10728:       |  |  |  |  |  >--|-> 02d7f73b                  remuw   a4,a5,a3
   1072c:       |  |  |  |  |  |  \-- db6d                      beqz    a4,1071e <main+0x7e>
   1072e:       |  |  |  |  |  |      1ae776d3                  fdiv.d  fa3,fa4,fa4
   10732:       |  |  |  |  |  |      2785                      addiw   a5,a5,1
   10734:       |  |  |  |  |  \----- fef49ae3                  bne     s1,a5,10728 <main+0x88>
   10738:       |  |  |  |  \-------> c0002773                  rdcycle a4
```

The branch predictor begins with the RISC-V ISA assumption: forwards
branches are not taken and backwards branches are not taken. This is to
handle the common cases of `if (ret == null) { abort(); }` for forward
branches and `for (...) { body; }` for backwards branches.

When the exception period is 2 (i.e. every OTHER FP instruction traps to
an FPE), then the branch predictor learns that 10724's `beq` is never
taken. This only happens after at least 3 iterations through the loop.
Now because the predictor has learned something, BOOM begins to
speculate (which itself is fine). BOOM will continue to speculate the
exceptional side of the `for`-loop, meaning it will put those 5
instructions into the ROB back to back several times (which is also
fine).

But, BOOM's ROB was never meant to handle exceptions raised by an
execute-stage functional unit. When the ROB informs the wider core about
the exception, it lowers the `r_xcpt_valid` and co. registers, which track
the current oldest registered exception. When we were speculating, we
created a tight loop of 5 instructions where once the first real `fdiv`
had its exception raised to the wider core, the second speculated `fdiv`
would be computed as the new "oldest" exception. But this was done
BEFORE the ROB began rolling-back. This prevented the ROB from cleaning
up the oldest registered exception.

Directly fixing this was easy, just clear `r_xcpt_valid` and co. when
rolling-back the core. But, that signal is also tied to the front-end's
`s0_valid` signal, which determines if the core should pull instructions
into the core. Simply clearing `r_xcpt_valid` this way prevented
`s0_valid` from ever going high again, which deadlocked the core, since
no new instructions could ever come in. This was fixed by using some
registered logic to prevent computing the new oldest registered
exception back-to-back with throwing it to the wider core.